Include image previews in the import image dialog
When importing an image that shares a palette with another image / tileset, the user may be curious to know what all images use the palette. The input window now has the ability to display "additional details", which right now is only used for sprites. But this would allow me to add other information to the options dialog in the future.