Implement Record fields
Records use the format |s={field}({value}={enum})
When the field (within the same array element) matches the given value, the record is treated as an enum of the given type. When the field is some other value, it's treated as just a number.

This is useful for evolutions, where the 'arg' is a different type depending on the evolution type.