Improve handling of broken pointers
If a pointer in a table points to the wrong type of data, we should recognize that and show that as an error pointer. For example, a pointer in a table that expects a compressed sprite, but gets an uncompressed one instead. In such cases, it should be possible to repoint from the bad data to new blank data.

If pointing to data that is the right type, but we don't point to the beginning of it, that should also allow the user to repoint to something fresh.