Only jump to moved data if the moved data was selected.
We want to show the user data that moves, but more than that, we don't want to take the user away from whatever they're currently looking at. When you change a palette, it may need to repoint a related tileset/tilemap, but we probably don't care. Just the notification is enough.