Implement write for 1-byte tilemaps
There is a second format for compressed tilemaps, where each tile gets only 1 byte instead of 2. This allows for a reduced number of tiles and no h/v flipping, but can save space.

This change implements the ability to import tilemaps with that format, as well as allowing the image editor to work with tilemaps of that format.