Populate Constant Reference
Move constants from default TOML files into the constantReference file. This makes it so that constants are condensed from 2010 lines spread between 9 files down to just 137 lines stored in a single file, but all the same information is available. This new format is much lighter on boilerplate, which makes it much faster to skim for constants.