Introduce new decorator format
SpriteDecorator wraps other formats, providing the extra behavior of having a sprite background. This can be used to draw a sprite behind a single cell or behind a group of cells.

To get the z-ordering right, the info from SpriteDecorator is read during a new RenderBackground method, which happens before RenderData.