Refactor ParentOffset to include a start/end pair instead of just a single value.
Before, ParentOffset was implemented as an int. If the value was negative, values are taken off from the start of the table. This is useful because some tables ignore pokemon 0 (named ??????) and start from Bulbasaur instead. If the value was positive, values are added to the end of the table. This is useful because some tables include extra elements at the end, like the "return" icon in the item image table, or the unown sprites in the pokemon sprite tables.

This is confusing, and is a problem if we have a table that needs elements removed from both the front and the back, or a table needs elements removed from the front and added to the back. We need a more flexible system.

The new implementation replaces ParentOffset with a pair of values, one for adjusting the front and one for adjusting the back. The old behavior remains: if a single number is given, it's assumed to be from the front if it's negative and from the back if it's positive. But now a pair of numbers can be given to be more explicit.