Implement table end-offset resize on Append
When appending elements to a table with a negative end-offset, clear the end offset and fill in the required number of elements. For example, if you add a new item to the items table, you want to know that the *same* new item is added to the item effects table.