Cut/Paste arrays should work for arrays with inner pointers
Inner Pointers were not previously correctly tracked during cut/paste operations. There needed to be some way to say that an address was attached not only to a name that does not exist, but to a specific index of such an array. The answer is to store the array offset in the bytes that are left behind, since the value of those bytes does not matter.