Initial Refactor for tileset limits
Introduce the concept of a 'max tiles' to the tileset format. Allow a tileset to know about the maximum number of tiles it's allowed to have.

By default, tilesets are limited to 1024 tiles. But some tilesets should have a reduced tile cap because of how they're used in the game. For example, Oak's background in the intro.