Introduce data.pokemon.type.length
Since the number of types is determined from code that most existing hacks don't change, that means that the length will be interpreted as the default 18 for almost all hacks.

Since hacks that expand types often customize the end of the type chart, make the match-requirement for recognizing table streams a bit less strict.

Update table-length-from-constant logic to be more fault tolerant and to take the constant multiplier into account. Also fix a bug: the ValueOffset should be subtracted off, not added on.

And most importantly: add constants and update tables involving types.