Improve loading edge case logic
Some loaded anchors depend on tables that are found, but may be the wrong length. If so, those tables will fail to load from metadata. Try a second time to load any missing tables after all the words have been loaded.

Improve Debug Conflict check to handle inner-anchors better

When increasing the size of a table after a constant update, make sure we keep hold of any pointers to inner elements.

When adding 1 to an existing SortedSpan, don't increment the collection size if the 1 is a duplicate.