Fix toml upgrade bugs
When upgrading TOML and a table length changes, observe the anchor instead of just observing the run.

Upgrade tests need to verify that no tables are missing.

Move items.stats to the top, since lots of other tables depend on it and it doesn't depend on other tables.

LzTilesets need to create LzTilesets when duplicated (matters during upgrade, since duplication is used to change the palette hint.)

Add DebugDictionary, a DictionaryDecorator (thanks Autoimplement) that makes it possible to observe when an element is removed from a dictionary.

Change the name of the "moveeffects" list to "moveeffectoptions". This is to avoid colliding with the previously expected name "moveeffects" in the tableReference file, which is the old name of scripts.moves.effects.

In the Upgrade Tests, verify that all the names in the fresh list exist in the upgraded list. It's ok if there's some cruft names in the upgrade list that don't appear in the fresh list.