Clean up refactor logic
I've proven that the new search routine works correctly for all 9 games for the .power field. As such, I've abstracted it out to a RefactorByteInTable function that can be called for any table, any field (that's a byte) to move to another location (must be a byte). I also wrote a draft for the code that will replace the ldrb command with a ldrh command, since I'll need to do that for the move effects.