First pass at ScrollRegion.AllowSingleTableMode
When this flag is active (single instance for the entire app, but propagated down to the ScrollRegion for use), the scroll region will respond differently to questions about its min/max scroll and the data length. This makes it possible to focus in on a single table that you're editing.
* This hides the "hex" part of the hex editor for simple use cases, and makes the right scrollbar actually useful for editing tables.
* It also makes it easier to focus in on individual tables and streams when jumping around with pointers.