Add map name popup themes
In Emerald, the box behind the mapname is themed. It looks like wood sometimes, or stone othertimes. Add in the palettes, background graphics, outline graphics, and the mapping data that connects each map name to its theme.

Note that the theme table starts out just like the map name table, but then skips 109 elements (the kanto maps) before putting on the last 16 elements. Because of this, the last 16 elements are mislabeled, and actually refer to the last 16 maps, not the first 16 Kanto maps.