Initial Pass at a MapListRun
Eventually, I'll need to have a way to tell how many maps are in each bank. Right now, I just have it hardcoded to work for an unmodified BPRE0. This class is untested, and mostly exists as a side project right now. But similar to how the overworld sprites table points to a bunch of overworld lists, each which knows how many elements long it is by doing some clever algorithm, I'd expect the list of map banks to each have a pointer to a map list.