Make QuickEditItems Async
Allow QuickEditItems to `await viewPort.UpdateProgress()` to temporarily get off the UI thread so that it can update the UI.
The new InlineDispatch class provides an awaitable object that can hook back into the dispatch interface. This allows code to be written using the IWorkDispatcher in a linear fashion.

This also temporarily disables the level-up move part of the move expansion, since I'm debugging part of the move expansion routine.