Improve QuickEditItem Loading Bar
In order to get the loading bar to work right for Edit operations, I needed a new implementation of Edit that can work with tasks. I need the method not to return until the edit is actually complete, so I can continue additional things after it. Also, I needed the edit to know that it's only a part of a range of the 100% instead of assuming that it gets to be all 100%.