Add Move Relearner Testing scirpt
This isn't a script meant to be shared. But it's useful for testing, so I'm including it as a resource. All it does is mess with the moveset of Bulbasaur/Torchic slightly and make the fat guy in the starting town into a free move relearner. This allows for easy testing of the level-up move code