Fix Array-with-TableStream-pointer expansion
When expanding and moving an ArrayRun that contains a pointer to table streams, the array needs to move, which means the table stream may temporarily have no pointers to the table. When the table re-adds itself on the other side, it's important that the table stream already know about the format of the parent table. For example, the length token may come from one of the parent element's fields.