Fix complete-edit-during-scroll behavior
When leaving a cell that's mid-edit, we usually try to complete the edit before clearing it. We now also do that when leaving the cell causes scrolling. This means that a few pieces of code that depended on the cells getting cleared automatically as part of the scroll (for example, extending arrays) needed to be updated so that the clear happens, then the try-complete code runs.