fix self-referential anchor recursion
Don't UpdateNewRunFromPointerFormat during AddPointerToAnchor if the pointer being added is within the existing run: doing so can cause so weird recursion.
Example: Emerald Advanced 0.7 has an event table at 748228 with a warp table pointer that leads to... 748228. In this case, we want to keep seeing it as an event table, not a warp table.