Fix UI notification from a parallel thread
(1) Find "welcome to the world of" (1 match, app jumps there). Is already formatted as text.
(2) Edit the text using the text tool on the left. Make the text shorter by deleting any one character.
(3) Find "one worth raising" (1 match, app jumps there). Is not formatted as text.
(4) Ctrl+D, T for Display as Text: crash!

This happens because 'display as text' has a lot of work to do, so it's sent off to be done in parallel. But if a pointer is found, adding that pointer to the collection counts as editing the token, which can cause the ChangeHistory to notify that certain bits of the menu need to be updated (such as the RunFile button). But those notifications can only happen on the UI thread, which means that the changes should only come from the UI thread.

Modify the code so that the change is logged and recorded, but not actually made until we're out of the parallel context.