add constants for ev caps
fixed a bug in R/S v1: when adding a constant, clear the existing format before adding it. This matters in case the constant was found in a location that contained auto-discovered text or sprite data that is in error.

Using ClearFormat can change the count of runs in an unexpected way, which makes the index we found unusable. So we need to use ObserveRunWritten like the other sections of this method do... which means we need to add a special "am I allowed to make changes" check into the ModelDelta code that handles tracking MatchedWord changes.