Improve Add-Fairy-Type script to allow for Fairy-Type TMs
In FR/LG, TMs get two images: one for shops and the PC etc, and one for the TM case. The TM case uses one sprite for all the TMs, but then uses a palette that's set based on the TMs type. These palettes need to be expanded to allow for a new Fairy-Type palette.