Improve sprite/palette bad compressed data behavior
When working with data that can be decoded but the end of the compressed data doesn't terminate correctly, we can still view the sprite/palette, but can't easily edit them, since the format is invalid.

When opening the image editor (or editing a palette in either the image editor or table tool), repoint invalid data as a new run and make it valid. This does a few things:
* Only the current pointer will be updated. Any other pointers to the bad data will remain the same, and the bad data will be left alone. If this is the only pointer to the bad data, it will _still_ be left alone, with only the new copy being used by the pointer. This ensures maximum compatibility, even if there's things going on that HMA doesn't know about.
* When coming from the sprite tool, nothing will be repointed upon opening the image editor, since that edit happens from the address and not from a pointer. When editing the image at this point, any required moves will update all known pointers at the same time.
* When any compressed data gets rewritten and moved, the user gets notified.