'Copy' shouldn't modify metadata
When running the Copy Command, don't modify the metadata by adding automatic names to the things being copied. We should still do that for cut, and the names should be added to the clipboard. But don't put the names in the actual rom. This should help reduce noise in the 'misc' anchor folder.