Parameterize FormatRunFactory
This allows us to have parameters from toml metadata passed in to the process of how we create formatted runs. In this change, we add the ability for trainer teams to _not_ scale the IV value from a number 0-255 to a number 0-31. This is important since some patches depend on using the entire byte, rather than just the top 5 bits.

Since the model is now in charge of the FormatRunFactory. the model can decide to use different types of factories for different roms. This lets us turn different features on/off between the pokemon games and other games. We may use this capability later.