Fix move expansion bug
Critical Hit Special Attack moves were doing more damage than they were supposed to after move expansion. After looking into it, this was because one instruction `mov   r0, #29` followed by `ldrsb r0, [r1, r0]`, was getting seen as an instruction pair that needed to be updated. But this is wrong for two reasons:
(1) A previous `mov r1` instruction should have stopped the algorithm before it reached this pair. The code has been fixed so that changes to the register from the original `add` instruction will now stop looking for this mov/ldrsb pair.
(2) The instruction we were trying to match was for moving #2 into a register, but we matched #29. Updated the code to expect the mov instruction to end with the constant, rather than just contain it.

I also updated the method documentation to explain how it handles a register source that starts with a !: in that case, it doesn't care if the target instruction contains the register, but it _does_ stop searching if the value in that register changes.

Tested in FireRed. This changes exactly one byte, which we verified through experimentation is the instruction that should not have been changed in the first place. So all is well now.