Add tests that explain how to deal with deeper trees.
Removing the leftmost node from a fully black subtree, the tree needs to balance itself using reds in the right subtree. The four cases are outer-red child and inner-red child for a red parent and a black parent. Each of the 4 cases results in a different balanced tree in the end, and none require further up-propagation.