update boxes menu palettes
Research shows that the addresses and dimensions graphics.menu.boxes.picker.tilemap  _are_ correct. However, HMA is unable to display the tilemap corrrectly, because the tilemap assumes that the tileset in question starts at tile 0x100 instead of tile 0x00. This is similar to how palettes can start at page 3 or 4 instead of having to start at page 0, and should probably be handled in a similar fashion, with the tileset being aware of its own offset.

HMA cannot handle this right now, so the tilemap is commented and untested.