Allow pokemon sprite import to import normal/shiny/front/back sets
There's a lot of resources available that use a 4x wide sprite to render the pokemon normal and shiny, front and back, all as a single image. People want to be able to import this all at once. Now they can.

As a side effect, pokemon will always show their normal palette as their first palette now, with the shiny palette always being second.

This only really works with single-frame sprites for now, so Ruby/FireRed, not Emerald.