Partial refactor for introducing SearchTree
Refactoring the code that cares about the index within the flat list so that I can more easily switch between the flat list implementation and the tree implementation for the set of runs. Not done refactoring yet, but this initial work gets the 'easy' bits out of the way so I can worry about the more complicated bits later.