Add limit to auto-length of tables
auto-length should be limited to the number of elements previously in the table, if a specific number is already known.
Note that this means that when a user updates from an older version with a shorter table to a newer version with a longer table, the additional elements table will only be found if the length is specified manually. Hence the change to `graphics.battle.background.sprites`.,