Add tables to tool in alphabetical order
Instead of adding tables to the tool as we find them, do a 2-pass system: in the first pass, add them to a sorted table. In the second pass, add those sorted tables to the tool.

This opens up options for further manual sorting via table groups in the metadata.