Pointers with anchors should show their own anchor, not the child anchor
HMA has a feature where you can see and edit anchors by clicking on a pointer to them, rather than having to go to the data. This is useful for stuff like editing the width/height and palette of a sprite.
If the pointer in question has an anchor, we should disable this feature and have the anchor editor show and edit the anchor for this address rather than the pointer's destination address.