Update Tilemap_UsesPalettePage_PaletteContainsPage
The logic to pull the data for this method used to actually load the entire model for Ruby/FireRed/Emerald. This is _very_ slow and made the tests start slowly, since enumerating them was such a cost.
The MemberData for this Theory now does all its work by looking only at tableReference, not actual models. This lets it load much faster.
Since it loads so fast, we can afford to check all 9 games that we support: R/S 1.0/1.1, FR/LG 1.0/1.1, and Emerald. So instead of only verifying R/FR/E, we can verify all 9.
Starting a test run or debug session is now much faster.