Fix metadata issue for extending child arrays from parent arrays
When we extend a child array from a parent array expanding, it's possible to knock out another child of the same parent, which then needs to be re-added. Doing so makes us lose the pointers of that child, which is a metadata bug.

Just like with the constant, make sure that updating the child table length needs to check that all associated pointers exist.