Add OffsetRenderViewModel
When a field in a table represents a graphical offset between a foreground and background element, users would like an easy way to see that without needing to run their game through an emulator.

This new addition to the table format allows for a 0-length format that takes parameters for background, foreground, and the relation between the two, allowing for x and y offsets to be applied between them.
In the table tool, this shows up as a composite image built from the background and foreground, and mouse drag can be used to change the relationship between the two.

Additionally, you can now use the arrow keys to increase/decrease values in numeric textboxes in the table tool. This is reversed for yoffset fields so that up/down move the sprite by a single pixel when pressed.