Update Hex selection when clicking in the table tool content
Do a goto whenever the table tool is clicked. Works for fields and drop-downs. Might add support for checkboxes in the future, depending on whether someone asks for it or not.

We specifically do _not_ update selection when clicking on stream content, because moving to the actual stream content can cause the table tool to change content.