Performance Improvement: table tool text editing
When you edit the pokemon names, another field in the table tool needs to update: the combobox drop-down list options for evolution species. There are 5 of these boxes.

Update the logic to be more choosy about what's being updated so that there's less changes going on and the app can update faster.