add newgame professor pokemon constants for R/S
We can't actually add these automatically to the metadata because of the way the code works: it's a mov/lsl pair, rather than just a flat mov or ldr. However, both routines that load this number have an extra `nop` at the end, which means there's enough space to safely adjust them to use `ldr` instructions instead.

Add a hma script to adjust the code for R/S and introduce the new constant. These are fairly small changes that only touch the two involved functions, so they should be safe to use on basically all R/S roms.