first pass at map tileset animation editor
map tileset animations consist of 3 parts:
(1) an `init` routine that sets a hook and some Max values
(2) a callback that runs every frame that decides whether or not to animate
(3) a set of uncompressed tiles to replace existing tiles in the tileset

For each replacement effect, you need to know how many frames are in the animation, how often to update to the next frame, how many tiles to replace, which tiles to replace, and the new tiles to insert. This is handled by a new table that's created specifically for this purpose, where the first field of the table is a pointer to a table of frames. This should make it a bit easier to understand, but a good user guide is still needed for this feature.

This doesn't work yet: there's an issue with the thumb code import step. But this is a first draft.