add protect success rate table
This table isn't terminated, which means it just _happens_ to drop to almost zero, since the metronome table is directly after it. But if the metronome table gets moved, then the rate grows back to full if the bytes after the table are FFFF. Adding this table makes it possible to fix the bug by expanding the table and adding `00 00` as the last entry, to make sure that Protect 'always fails on the 5th attempt' or however many entries you want to add.