use a threadsafe dictionary for address/anchor mapping
In rare cases, you can get a race condition where we try to turn the address-for-anchor keys into a list while also adding a new address-for-anchor key. This multi-threaded work can crash the app.

Use a thread-locking version of version of the dictionary to prevent Keys from being accessed at the same time as Add.