don't auto-change palettes when 8x8 editing a 16 color sprite
If a single sprite only supports 16 colors but has multiple palettes, you can cycle through the available palettes while editing it. But editing a 8x8 tile has special logic that causes it to switch the palette associated with a tile... which doesn't work in this case. So the 8x8 special logic should be skipped if its a 16-color image.