improve paste logic
when pasting text, always complete the current cell by putting an extra space on the end. This might not be quite what we want, as we may want to paste something and then be in the middle of a cell edit... but more often, we want to paste multiple cells, and the last cell doesn't commit correctly if we didn't copy a space.