fix table expansion anchor glitch
Expanding a table where the last element has a named anchor adds another pointer to that anchor. Some format types, such as palettes, do a ClearFormat when adding the new run. This is reasonable, but causes us to lose any custom name attached to those elements, which update their pointers to point to null and adds the name to the unmapped anchor names list.

In this case, we need to track the name before we clear the format, so that when we add the new run, we can add the name as well.